Upload containing the working version of the save_qualifying_loans function and updated save_csv.py file and function

+++ loan_qualifier_app/qualifier/utils/save_csv.py
@@ -1,7 +1,13 @@
 import csv
+import sys
 
-def save_csv(qualifying_loans):
-    csvpath = Path("qualifying_loans.csv")
+from pathlib import Path
+
+def save_csv(qualifying_loans, csvpath):
+    csvpath = Path(csvpath)
+    
+    if not csvpath.exists():
+        sys.exit(f"Oops! Can't find this path: {csvpath}")
     
     with open(csvpath, 'w', newline='') as csvfile:
         csvwriter = csv.writer(csvfile)
